[HACK] drivers: bq2597x: Comment out checks for device awake state
bq_charge_awake returns -16 breaking deep sleep while charging. I tried to trace the issue but was unable to find out the root cause.
Disabling this check allows the device to suspend during charging.

Please note that this is by no means a proper fix. Will be reverted when a proper fix is found.

Signed-off-by: taalojarvi <sreedevan05@gmail.com>